Refactoring

+++ 1Set/Concrete/ExercisesRepository.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Set.Abstract;
-using Set.Models;
-using SQLite.Net.Async;
+using Set.Entities;
+using SQLite;
 
 namespace Set.Concrete
 {
@@ -16,10 +16,15 @@ namespace Set.Concrete
 		public override async Task<int> DeleteAsync(int id)
 		{
 			using (await Mutex.LockAsync ().ConfigureAwait (false))
-			{				
-				await _connection.ExecuteAsync ("DELETE FROM RoutineDays WHERE ExerciseId = ?", id);
-				await _connection.ExecuteAsync ("DELETE FROM Workouts WHERE ExerciseId = ?", id);
-				return await _connection.DeleteAsync<Exercise> (id);			
+			{
+                var item = FindAsync(id);
+                if (item != null)
+                {
+                    await _connection.ExecuteAsync("DELETE FROM RoutineDays WHERE ExerciseId = ?", id);
+                    await _connection.ExecuteAsync("DELETE FROM Workouts WHERE ExerciseId = ?", id);
+                    return await _connection.DeleteAsync(item);
+                }
+                return 0;
 			}
 		}
